wnt adopted version of runSalome and killSalome scripts

+++ bin/killSalomeWithPort.py
@@ -22,7 +22,39 @@ def getPiDict(port,appname='salome',full=True):
         filedict = os.getenv("HOME") + '/' + filedict
     return filedict
 
+def appliCleanOmniOrbConfig(port):
+    """
+    remove omniorb config files related to the port in SALOME application:
+    - ${HOME}/${APPLI}/.omniORB_${HOSTNAME}_${NSPORT}.cfg
+    - ${HOME}/${APPLI}/.omniORB_${HOSTNAME}_last.cfg
+    the last is removed only if the link points to the first file.
+    """
+    from Utils_Identity import getShortHostName
+    appli=os.environ.get("APPLI")
+    if appli is None:
+        #Run outside application context
+        pass
+    else:
+        home = os.environ['HOME']
+        home='%s/%s'%(home,appli)
+        hostname=getShortHostName()
+        omniorb_config = '%s/.omniORB_%s_%s.cfg'%(home,hostname, str(port))
+        last_running_config = '%s/.omniORB_%s_last.cfg'%(home, hostname)
+        if os.access(last_running_config,os.F_OK):
+            pointedPath = os.readlink(last_running_config)
+            if pointedPath[0] != '/':
+                pointedPath=os.path.join(os.path.dirname(last_running_config), pointedPath)
+            if pointedPath == omniorb_config:
+                os.unlink(last_running_config)
+                pass
+            pass
+        if os.access(omniorb_config,os.F_OK):
+            os.remove(omniorb_config)
+            pass
+        pass
+
 ########## kills all salome processes with the given port ##########
+
 def killMyPort(port):
     filedict=getPiDict(port)
     found = 0
@@ -30,11 +62,30 @@ def killMyPort(port):
         fpid=open(filedict, 'r')
         found = 1
     except:
-        print "le fichier %s des process SALOME n'est pas accessible"% filedict
+        print "file %s giving SALOME process id is not readable"% filedict
         pass
         
     if found:
-        a = os.system("pid=`ps -eo pid,command | egrep \"[0-9] omniNames -start "+str(port)+"\" | sed -e \"s%[^0-9]*\([0-9]*\) .*%\\1%g\"`; kill -9 $pid >& /dev/null")
+        #cmd = 'pid=`ps -eo pid,command | egrep "[0-9] omniNames -start '+str(port)+'"` ; echo $pid > /tmp/logs/'+os.getenv('USER')+"/_"+port+'_Pid_omniNames.log'
+        #a = os.system(cmd)
+        try:
+            fpidomniNames=open('/tmp/logs/'+os.getenv('USER')+"/_"+port+'_Pid_omniNames.log')
+            prc = fpidomniNames.read()
+            fpidomniNames.close()
+            if prc != None :
+                for field in prc.split(" ") :
+                    if field == "omniNames" :
+                        if pidfield != "egrep" :
+                            print 'stop process '+pidfield+' : omniNames'
+	                    if sys.platform == "win32":
+                              import win32pm
+                              win32pm.killpid(int(pidfield),0)
+                            else:
+                              os.system('kill -9 '+pidfield)
+                    pidfield = field
+        except:
+            pass
+        
         try:
             process_ids=pickle.load(fpid)
             fpid.close()
@@ -43,21 +94,32 @@ def killMyPort(port):
                 for pid, cmd in process_id.items():
                     print "stop process %s : %s"% (pid, cmd[0])
                     try:
-                        os.kill(int(pid),signal.SIGKILL)
+	                if sys.platform == "win32":
+                          import win32pm
+                          win32pm.killpid(int(pid),0)
+                        else:
+                          os.kill(int(pid),signal.SIGKILL)
                     except:
-                        print "  ------------------ process %s : %s inexistant"% (pid, cmd[0])
+                        print "  ------------------ process %s : %s not found"% (pid, cmd[0])
                         pass
                 pass
         except:
             pass
+        
         os.remove(filedict)
-        pid = commands.getoutput("ps -eo pid,command | egrep \"[0-9] omniNames -start "+str(port)+"\" | sed -e \"s%[^0-9]*\([0-9]*\) .*%\\1%g\"")
-	a = ""
+        cmd='ps -eo pid,command | egrep "[0-9] omniNames -start '+str(port)+'" | sed -e "s%[^0-9]*\([0-9]*\) .*%\\1%g"'
+        pid = commands.getoutput(cmd)
+        a = ""
         while pid != "" and len(a.split(" ")) < 2:
             a = commands.getoutput("pid=`ps -eo pid,command | egrep \"[0-9] omniNames -start "+str(port)+"\" | sed -e \"s%[^0-9]*\([0-9]*\) .*%\\1%g\"`; kill -9 $pid")
             pid = commands.getoutput("ps -eo pid,command | egrep \"[0-9] omniNames -start "+str(port)+"\" | sed -e \"s%[^0-9]*\([0-9]*\) .*%\\1%g\"")
 	    print pid
-              
+            
+        pass
+
+    appliCleanOmniOrbConfig(port)
+    pass
+            
 
 if __name__ == "__main__":
     for port in sys.argv[1:]:
